refactor: do not expose `policyProposalMaxExecutables`
